EE-9765 After code review: Renamed AliasSurnameCombinations to AliasCombinations and functions class renamed similarly. Moved tests for AliasCombinationsFunctions into a separate test class per function
