+++ src/test/java/uk/gov/digital/ho/pttg/application/namematching/candidates/AliasCombinationsFunctionsNonAliasFirstLastTest.java
@@ -1,0 +1,72 @@
+package uk.gov.digital.ho.pttg.application.namematching.candidates;
+
+import org.junit.Test;
+import uk.gov.digital.ho.pttg.application.namematching.CandidateName;
+
+import java.util.List;
+
+import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
+import static java.util.Collections.singletonList;
+import static org.assertj.core.api.Assertions.assertThat;
+import static uk.gov.digital.ho.pttg.application.namematching.candidates.AliasCombinationsFunctions.nonAliasFirstNamesAndLastNameCombinations;
+
+public class AliasCombinationsFunctionsNonAliasFirstLastTest {
+    @Test
+    public void shouldReturnEmptyListForEmptyInputs() {
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(emptyList(), emptyList())).isEqualTo(emptyList());
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(singletonList("somename"), emptyList())).isEqualTo(emptyList());
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(emptyList(), singletonList("somename"))).isEqualTo(emptyList());
+    }
+
+    @Test
+    public void combinationsForOneFirstNameOneLastName() {
+        List<String> lastNames = singletonList("Smith");
+        List<String> allNames = asList("John", "Smith");
+        List<CandidateName> expected = singletonList(new CandidateName("John", "Smith"));
+
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(allNames, lastNames)).isEqualTo(expected);
+    }
+
+    @Test
+    public void combinationsForTwoFirstNameOneLastName() {
+        List<String> lastNames = singletonList("Smith");
+        List<String> allNames = asList("John", "David", "Smith");
+        List<CandidateName> expected = asList(
+                new CandidateName("John", "Smith"),
+                new CandidateName("David", "Smith")
+        );
+
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(allNames, lastNames)).isEqualTo(expected);
+    }
+
+    @Test
+    public void combinationsForOneFirstNameTwoLastNames() {
+        List<String> lastNames = asList("Smith", "Evans");
+        List<String> allNames = asList("John", "Smith", "Evans");
+        List<CandidateName> expected = asList(
+                new CandidateName("John", "Evans"),
+                new CandidateName("Smith", "Evans"),
+                new CandidateName("John", "Smith"),
+                new CandidateName("Evans", "Smith")
+        );
+
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(allNames, lastNames)).isEqualTo(expected);
+    }
+
+    @Test
+    public void combinationsForTwoFirstNamesTwoLastName() {
+        List<String> lastNames = asList("Smith", "Evans");
+        List<String> allNames = asList("John", "David", "Smith", "Evans");
+        List<CandidateName> expected = asList(
+                new CandidateName("John", "Evans"),
+                new CandidateName("David", "Evans"),
+                new CandidateName("Smith", "Evans"),
+                new CandidateName("John", "Smith"),
+                new CandidateName("David", "Smith"),
+                new CandidateName("Evans", "Smith")
+        );
+
+        assertThat(nonAliasFirstNamesAndLastNameCombinations(allNames, lastNames)).isEqualTo(expected);
+    }
+}
